48_Transaction_controller_and_endpoints
- created the transaction controller which handles requests like topup,
list of transactions and send
- added service methods that handle the logic

+++ src/main/java/com/jimaio/vanir/domain/Account.java
@@ -19,8 +19,6 @@ import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 import javax.persistence.Transient;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
 import lombok.Data;
 
 @Entity
@@ -40,7 +38,7 @@ public class Account implements Serializable {
 	protected User user;
 	
 	@Column(name="BALANCE")
-	protected Integer balance;
+	protected Double balance;
 	
 	@JoinColumn(name="CURRENCY_ID", nullable=false, updatable=true, insertable=true)
 	protected Currency currency;
